bypass login if valid session, use ora to display check
