+++ packages/sm-init/src/steps/index.ts
@@ -1,3 +1,4 @@
-export { installSm } from './install-sm'
-export { validatePkg } from './validate-pkg'
-export { addScriptToPkg } from './add-script-to-pkg'
+export { installSm } from "./install-sm";
+export { validatePkg } from "./validate-pkg";
+export { addScriptToPkg } from "./add-script-to-pkg";
+export { loginOrBypass } from "./loginOrBypass";
